Remove Container in between
