+++ front/app/components/ConsentManager/Container.tsx
@@ -1,6 +1,4 @@
-import React, { FormEvent, useState } from 'react';
-
-import useObserveEvent from 'hooks/useObserveEvent';
+import React from 'react';
 
 import Banner from './Banner';
 import { TCategory } from './destinations';
@@ -10,54 +8,35 @@ import { CategorizedDestinations, IPreferences } from './typings';
 interface Props {
   preferences: IPreferences;
   categorizedDestinations: CategorizedDestinations;
-  isConsentRequired: boolean;
-  onToggleModal: (opened: boolean) => void;
   updatePreference: (category: TCategory, value: boolean) => void;
-  resetPreferences: () => void;
   accept: () => void;
   reject: () => void;
-  saveConsent: () => void;
+  openDialog: () => void;
+  closeDialog: () => void;
+  isDialogOpen: boolean;
+  handleCancel: () => void;
+  handleSave: () => void;
 }
 
 const Container = ({
   preferences,
   categorizedDestinations,
-  isConsentRequired,
-  onToggleModal,
   updatePreference,
-  resetPreferences,
   accept,
   reject,
-  saveConsent,
+  openDialog,
+  closeDialog,
+  isDialogOpen,
+  handleCancel,
+  handleSave,
 }: Props) => {
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-
-  const openDialog = () => {
-    onToggleModal(false);
-    setIsDialogOpen(true);
-  };
-
-  const closeDialog = () => {
-    onToggleModal(true);
-    setIsDialogOpen(false);
-  };
-
-  useObserveEvent('openConsentManager', openDialog);
-
-  const handleSave = (e: FormEvent) => {
-    e.preventDefault();
-
-    setIsDialogOpen(false);
-    saveConsent();
-  };
-
-  const handleCancel = () => {
-    resetPreferences();
-    setIsDialogOpen(false);
-  };
-
   return (
     <>
+      <Banner
+        onAccept={accept}
+        onChangePreferences={openDialog}
+        onClose={reject}
+      />
       <PreferencesModal
         opened={isDialogOpen}
         categorizedDestinations={categorizedDestinations}
@@ -67,14 +46,6 @@ const Container = ({
         onClose={closeDialog}
         updatePreference={updatePreference}
       />
-
-      {isConsentRequired && (
-        <Banner
-          onAccept={accept}
-          onChangePreferences={openDialog}
-          onClose={reject}
-        />
-      )}
     </>
   );
 };
